Change variable parameter type for the 'id' field to int for all daos and classes

+++ src/main/java/dunn/ted/java/instateam/dao/CollaboratorDao.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public interface CollaboratorDao {
     List<Collaborator> findAll();
-    Collaborator findById(Long id);
+    Collaborator findById(int id);
     void save(Collaborator collaborator);
     void delete(Collaborator collaborator);
 }
