+++ src/main/java/dunn/ted/java/instateam/dao/RoleDao.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public interface RoleDao {
     List<Role> findAll();
-    Role findById(Long id);
+    Role findById(int id);
     void save(Role role);
     void delete(Role role);
 }
